Remove old service code
